flod_tasks/requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.18
celery==3.1.13
kombu==3.0.21
pytz==2014.7
requests==2.2.1
suds-jurko==0.6
